+++ packages/react-doctor/package.json
@@ -82,6 +82,7 @@
   },
   "devDependencies": {
     "@types/prompts": "^2.4.9",
+    "@typescript-eslint/types": "^8.59.3",
     "eslint-plugin-react-hooks": "^7.1.1",
     "eslint-plugin-react-you-might-not-need-an-effect": "^0.10.1"
   },
