+++ packages/react-doctor/src/plugin/types.ts
@@ -1,33 +1,9 @@
-interface ReportDescriptor {
-  node: EsTreeNode;
-  message: string;
-}
-
-export interface RuleContext {
-  report: (descriptor: ReportDescriptor) => void;
-  getFilename?: () => string;
-}
-
-export interface RuleVisitors {
-  [selector: string]: ((node: EsTreeNode) => void) | (() => void);
-}
-
-export interface Rule {
-  create: (context: RuleContext) => RuleVisitors;
-}
-
-export interface RulePlugin {
-  meta: { name: string };
-  rules: Record<string, Rule>;
-}
-
-export interface EsTreeNode {
-  type: string;
-  [key: string]: any;
-}
-
-export interface ParsedRgb {
-  red: number;
-  green: number;
-  blue: number;
-}
+export type {
+  EsTreeNode,
+  ParsedRgb,
+  ReportDescriptor,
+  Rule,
+  RuleContext,
+  RulePlugin,
+  RuleVisitors,
+} from "./utils/index.js";
