+++ packages/react-doctor/src/plugin/utils/component-binding-stack-tracker-callbacks.ts
@@ -1,0 +1,5 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export interface ComponentBindingStackTrackerCallbacks {
+  onVariableDeclarator?: (node: EsTreeNode) => void;
+}
