+++ packages/react-doctor/src/plugin/utils/component-binding-stack-tracker.ts
@@ -1,0 +1,8 @@
+import type { RuleVisitors } from "./rule-visitors.js";
+
+export interface ComponentBindingStackTracker {
+  isInsideComponent: () => boolean;
+  isBoundName: (name: string) => boolean;
+  addBindingToCurrentFrame: (name: string) => void;
+  visitors: RuleVisitors;
+}
