+++ packages/react-doctor/src/plugin/utils/component-prop-stack-tracker-callbacks.ts
@@ -1,0 +1,5 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export interface ComponentPropStackTrackerCallbacks {
+  onComponentEnter?: (componentBody: EsTreeNode | undefined) => void;
+}
