+++ packages/react-doctor/src/plugin/utils/component-prop-stack-tracker.ts
@@ -1,0 +1,7 @@
+import type { RuleVisitors } from "./rule-visitors.js";
+
+export interface ComponentPropStackTracker {
+  isPropName: (name: string) => boolean;
+  getCurrentPropNames: () => Set<string>;
+  visitors: RuleVisitors;
+}
