+++ packages/react-doctor/src/plugin/utils/contains-fetch-call.ts
@@ -1,0 +1,21 @@
+import { FETCH_CALLEE_NAMES, FETCH_MEMBER_OBJECTS } from "../constants.js";
+import type { EsTreeNode } from "./es-tree-node.js";
+import { walkAst } from "./walk-ast.js";
+
+export const containsFetchCall = (node: EsTreeNode): boolean => {
+  let didFindFetchCall = false;
+  walkAst(node, (child) => {
+    if (didFindFetchCall || child.type !== "CallExpression") return;
+    if (child.callee?.type === "Identifier" && FETCH_CALLEE_NAMES.has(child.callee.name)) {
+      didFindFetchCall = true;
+    }
+    if (
+      child.callee?.type === "MemberExpression" &&
+      child.callee.object?.type === "Identifier" &&
+      FETCH_MEMBER_OBJECTS.has(child.callee.object.name)
+    ) {
+      didFindFetchCall = true;
+    }
+  });
+  return didFindFetchCall;
+};
