+++ packages/react-doctor/src/plugin/utils/count-set-state-calls.ts
@@ -1,0 +1,11 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { isSetterCall } from "./is-setter-call.js";
+import { walkAst } from "./walk-ast.js";
+
+export const countSetStateCalls = (node: EsTreeNode): number => {
+  let setStateCallCount = 0;
+  walkAst(node, (child) => {
+    if (isSetterCall(child)) setStateCallCount++;
+  });
+  return setStateCallCount;
+};
