+++ packages/react-doctor/src/plugin/utils/create-loop-aware-visitors.ts
@@ -1,0 +1,30 @@
+import { LOOP_TYPES } from "../constants.js";
+import type { EsTreeNode } from "./es-tree-node.js";
+import type { RuleVisitors } from "./rule-visitors.js";
+
+export const createLoopAwareVisitors = (
+  innerVisitors: Record<string, (node: EsTreeNode) => void>,
+): RuleVisitors => {
+  let loopDepth = 0;
+  const incrementLoopDepth = (): void => {
+    loopDepth++;
+  };
+  const decrementLoopDepth = (): void => {
+    loopDepth--;
+  };
+
+  const visitors: RuleVisitors = {};
+
+  for (const loopType of LOOP_TYPES) {
+    visitors[loopType] = incrementLoopDepth;
+    visitors[`${loopType}:exit`] = decrementLoopDepth;
+  }
+
+  for (const [nodeType, handler] of Object.entries(innerVisitors)) {
+    visitors[nodeType] = (node: EsTreeNode) => {
+      if (loopDepth > 0) handler(node);
+    };
+  }
+
+  return visitors;
+};
