+++ packages/react-doctor/src/plugin/utils/define-rule.ts
@@ -1,0 +1,9 @@
+import type { Rule } from "./rule.js";
+
+export interface DefineRule {
+  (rule: Rule): Rule;
+  <RuleDefinition>(rule: RuleDefinition): RuleDefinition;
+}
+
+export const defineRule: DefineRule = <RuleDefinition>(rule: RuleDefinition): RuleDefinition =>
+  rule;
