+++ packages/react-doctor/src/plugin/utils/find-jsx-attribute.ts
@@ -1,0 +1,12 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export const findJsxAttribute = (
+  attributes: EsTreeNode[],
+  attributeName: string,
+): EsTreeNode | undefined =>
+  attributes?.find(
+    (attribute: EsTreeNode) =>
+      attribute.type === "JSXAttribute" &&
+      attribute.name?.type === "JSXIdentifier" &&
+      attribute.name.name === attributeName,
+  );
