+++ packages/react-doctor/src/plugin/utils/find-side-effect.ts
@@ -1,0 +1,33 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { isCookiesOrHeadersCall } from "./is-cookies-or-headers-call.js";
+import { isMutatingDbCall } from "./is-mutating-db-call.js";
+import { isMutatingFetchCall } from "./is-mutating-fetch-call.js";
+import { isMutatingMethodProperty } from "./is-mutating-method-property.js";
+import { walkAst } from "./walk-ast.js";
+
+export const findSideEffect = (node: EsTreeNode): string | null => {
+  let sideEffectDescription: string | null = null;
+  walkAst(node, (child: EsTreeNode) => {
+    if (sideEffectDescription) return;
+    if (isCookiesOrHeadersCall(child, "cookies")) {
+      const methodName = child.callee.property.name;
+      sideEffectDescription = `cookies().${methodName}()`;
+    } else if (isCookiesOrHeadersCall(child, "headers")) {
+      const methodName = child.callee.property.name;
+      sideEffectDescription = `headers().${methodName}()`;
+    } else if (isMutatingFetchCall(child)) {
+      // HACK: re-use the EXACT predicate `isMutatingFetchCall` already
+      // matched on so we can't pick a non-Literal duplicate `method:`
+      // entry by mistake (a looser `key.name === "method"` predicate
+      // would).
+      const methodProperty = child.arguments[1].properties.find(isMutatingMethodProperty);
+      sideEffectDescription = `fetch() with method ${methodProperty.value.value}`;
+    } else if (isMutatingDbCall(child)) {
+      const methodName = child.callee.property.name;
+      const objectName =
+        child.callee.object?.type === "Identifier" ? child.callee.object.name : null;
+      sideEffectDescription = objectName ? `${objectName}.${methodName}()` : `.${methodName}()`;
+    }
+  });
+  return sideEffectDescription;
+};
