+++ packages/react-doctor/src/plugin/utils/get-callback-statements.ts
@@ -1,0 +1,8 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export const getCallbackStatements = (callback: EsTreeNode): EsTreeNode[] => {
+  if (callback.body?.type === "BlockStatement") {
+    return callback.body.body ?? [];
+  }
+  return callback.body ? [callback.body] : [];
+};
