+++ packages/react-doctor/src/plugin/utils/has-directive.ts
@@ -1,0 +1,11 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export const hasDirective = (programNode: EsTreeNode, directive: string): boolean =>
+  Boolean(
+    programNode.body?.some(
+      (statement: EsTreeNode) =>
+        statement.type === "ExpressionStatement" &&
+        statement.expression?.type === "Literal" &&
+        statement.expression.value === directive,
+    ),
+  );
