+++ packages/react-doctor/src/plugin/utils/has-jsx-attribute.ts
@@ -1,0 +1,5 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { findJsxAttribute } from "./find-jsx-attribute.js";
+
+export const hasJsxAttribute = (attributes: EsTreeNode[], attributeName: string): boolean =>
+  Boolean(findJsxAttribute(attributes, attributeName));
