+++ packages/react-doctor/src/plugin/utils/is-ast-node.ts
@@ -1,0 +1,7 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { hasTypeProperty } from "./has-type-property.js";
+
+export const isAstNode = (value: unknown): value is EsTreeNode => {
+  if (!hasTypeProperty(value)) return false;
+  return typeof value.type === "string";
+};
