+++ packages/react-doctor/src/plugin/utils/is-component-assignment.ts
@@ -1,0 +1,9 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { isUppercaseName } from "./is-uppercase-name.js";
+
+export const isComponentAssignment = (node: EsTreeNode): boolean =>
+  node.type === "VariableDeclarator" &&
+  node.id?.type === "Identifier" &&
+  isUppercaseName(node.id.name) &&
+  Boolean(node.init) &&
+  (node.init.type === "ArrowFunctionExpression" || node.init.type === "FunctionExpression");
