+++ packages/react-doctor/src/plugin/utils/is-hook-call.ts
@@ -1,0 +1,9 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { getCalleeName } from "./get-callee-name.js";
+
+export const isHookCall = (node: EsTreeNode, hookName: string | Set<string>): boolean => {
+  if (node.type !== "CallExpression") return false;
+  const calleeName = getCalleeName(node);
+  if (!calleeName) return false;
+  return typeof hookName === "string" ? calleeName === hookName : hookName.has(calleeName);
+};
