+++ packages/react-doctor/src/plugin/utils/is-setter-identifier.ts
@@ -1,0 +1,3 @@
+import { SETTER_PATTERN } from "../constants.js";
+
+export const isSetterIdentifier = (name: string): boolean => SETTER_PATTERN.test(name);
