+++ packages/react-doctor/src/plugin/utils/is-simple-expression.ts
@@ -1,0 +1,25 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export const isSimpleExpression = (node: EsTreeNode | null): boolean => {
+  if (!node) return false;
+  switch (node.type) {
+    case "Identifier":
+    case "Literal":
+    case "TemplateLiteral":
+      return true;
+    case "BinaryExpression":
+      return isSimpleExpression(node.left) && isSimpleExpression(node.right);
+    case "UnaryExpression":
+      return isSimpleExpression(node.argument);
+    case "MemberExpression":
+      return !node.computed && isSimpleExpression(node.object);
+    case "ConditionalExpression":
+      return (
+        isSimpleExpression(node.test) &&
+        isSimpleExpression(node.consequent) &&
+        isSimpleExpression(node.alternate)
+      );
+    default:
+      return false;
+  }
+};
