+++ packages/react-doctor/src/plugin/utils/is-uppercase-name.ts
@@ -1,0 +1,3 @@
+import { UPPERCASE_PATTERN } from "../constants.js";
+
+export const isUppercaseName = (name: string): boolean => UPPERCASE_PATTERN.test(name);
