+++ packages/react-doctor/src/plugin/utils/parsed-rgb.ts
@@ -1,0 +1,5 @@
+export interface ParsedRgb {
+  red: number;
+  green: number;
+  blue: number;
+}
