+++ packages/react-doctor/src/plugin/utils/report-descriptor.ts
@@ -1,0 +1,6 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export interface ReportDescriptor {
+  node: EsTreeNode;
+  message: string;
+}
