+++ packages/react-doctor/src/plugin/utils/rule-context.ts
@@ -1,0 +1,6 @@
+import type { ReportDescriptor } from "./report-descriptor.js";
+
+export interface RuleContext {
+  report: (descriptor: ReportDescriptor) => void;
+  getFilename?: () => string;
+}
