+++ packages/react-doctor/src/plugin/utils/rule-example.ts
@@ -1,0 +1,4 @@
+export interface RuleExample {
+  before: string;
+  after: string;
+}
