+++ packages/react-doctor/src/plugin/utils/rule-plugin.ts
@@ -1,0 +1,6 @@
+import type { Rule } from "./rule.js";
+
+export interface RulePlugin {
+  meta: { name: string };
+  rules: Record<string, Rule>;
+}
