+++ packages/react-doctor/src/plugin/utils/rule-visitors.ts
@@ -1,0 +1,5 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+
+export interface RuleVisitors {
+  [selector: string]: ((node: EsTreeNode) => void) | (() => void);
+}
