+++ packages/react-doctor/src/plugin/utils/value-with-type.ts
@@ -1,0 +1,3 @@
+export interface ValueWithType {
+  type?: unknown;
+}
