+++ packages/react-doctor/src/plugin/utils/walk-ast.ts
@@ -1,0 +1,23 @@
+import type { EsTreeNode } from "./es-tree-node.js";
+import { isAstNode } from "./is-ast-node.js";
+
+// HACK: AST is acyclic except for `parent` back-references, which we skip.
+// Visitors may return `false` to prune the subtree below `node` (e.g. to
+// stop walking into nested functions when collecting `await` expressions
+// for the enclosing function only). Returning anything else (including
+// `undefined`, the natural value of statements) continues the walk.
+export const walkAst = (node: EsTreeNode, visitor: (child: EsTreeNode) => boolean | void): void => {
+  if (!node || typeof node !== "object") return;
+  if (visitor(node) === false) return;
+  for (const key of Object.keys(node)) {
+    if (key === "parent") continue;
+    const child = node[key];
+    if (Array.isArray(child)) {
+      for (const item of child) {
+        if (isAstNode(item)) walkAst(item, visitor);
+      }
+    } else if (isAstNode(child)) {
+      walkAst(child, visitor);
+    }
+  }
+};
